@@ -4,10 +4,10 @@ import android.app.Activity
 import android.content.Context
 import android.content.Intent
 import android.os.Bundle
-import android.support.design.widget.FloatingActionButton
-import android.support.v7.app.AppCompatActivity
-import android.support.v7.widget.GridLayoutManager
-import android.support.v7.widget.RecyclerView
+import com.google.android.material.floatingactionbutton.FloatingActionButton
+import androidx.appcompat.app.AppCompatActivity
+import androidx.recyclerview.widget.GridLayoutManager
+import androidx.recyclerview.widget.RecyclerView
 import android.view.View
 import com.rajat.alldemoapp.app_manager.AppsManager
 import com.rajat.alldemoapp.app_manager.UninstallAppsAdapter
@@ -73,6 +73,6 @@ class MainActivity : AppCompatActivity() {
 
     override fun onResume() {
         super.onResume()
-        resetAdapter(mActivity as MainActivity, this!!.mRecyclerView!!)
+        resetAdapter(mActivity as MainActivity, this.mRecyclerView!!)
     }
 }
